chore(dashboard): cleanup
